Added better ActiveModel support
Sniffs to see if ActiveModel is defined. If so, includes the custom validator
into ActiveModel::Validations::HelperMethods (for Rails 3). If not, extends
ActiveRecord::Base (for Rails 2.x) -- maintaining backwards compatibility.

This allows the custom email validator to be used in environments where
ActiveRecord is not loaded, but that still rely on ActiveModel.